Jacmazon_ECommerce/Repositories/CRUDRepository.cs
using Microsoft.EntityFrameworkCore;
using System.Linq.Expressions;
using Jacmazon_ECommerce.Data;
using System.Reflection;
using NuGet.Protocol;

namespace Jacmazon_ECommerce.Repositories
{
    public class CRUDRepository<TEntity, TContext> : ICRUDRepository<TEntity> 
        where TEntity : class
        where TContext : DbContext
    {
        private readonly TContext _context;
        private readonly DbSet<TEntity> _dbSet;

        public CRUDRepository(TContext context)
        {
            _context = context;
            _dbSet = _context.Set<TEntity>();
        }

        public async Task<bool> CreateAsync(TEntity entity)
        {
            await _dbSet.AddAsync(entity);
            int count =  await _context.SaveChangesAsync();
            return count > 0;
        }

        public async Task<int> UpdateAsync(TEntity entity)
        {
            _dbSet.Update(entity);
            return await _context.SaveChangesAsync();
        }

        public async Task<int> DeleteAsync(TEntity entity)
        {
            _dbSet.Remove(entity);
            return await _context.SaveChangesAsync();
        }

        public async Task<TEntity?> FindByIdAsync(int id)
        {
            var a = await _context.FindAsync(typeof(TEntity), 1);
            return await _dbSet.FindAsync(id);
        }

        public async Task<IEnumerable<TEntity>> FindAsync(Expression<Func<TEntity, bool>> expression)
        {
            return await _dbSet.Where(expression).ToListAsync();
        }
    }
}
